Projeto-01-Ingestão/pyproject.toml
[project]
name = "projeto01-injestao"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "python-dotenv>=1.2.2",
    "requests>=2.33.1",
]
